removed react-juce and working on fixing warnings

+++ PluginName/Source/PluginNameGUI.h
@@ -13,8 +13,8 @@
 //==============================================================================
 class PluginNameGUI {
 public:
-    static constexpr float ratio = 1.618; // golden ratio
-    static constexpr int height = 400,
+    static constexpr float ratio = 1.618f; // golden ratio
+    static constexpr float height = 400,
                          width  = height * ratio;
     
     //==============================================================================
@@ -35,7 +35,7 @@ public:
 class MenuPanelGUI {
 public:
     //==============================================================================
-    static constexpr int width  = PluginNameGUI::width * 0.13f,
+    static constexpr float width  = PluginNameGUI::width * 0.13f,
                          height = PluginNameGUI::height;
 };
 
@@ -43,20 +43,20 @@ public:
 class PresetPanelGUI {
 public:
     //==============================================================================
-    static constexpr int width  = PluginNameGUI::width - MenuPanelGUI::width,
+    static constexpr float width  = PluginNameGUI::width - MenuPanelGUI::width,
                          height = PluginNameGUI::height * 0.09f;
     
     //==============================================================================
-    static constexpr int undo_redo_width  = height * 0.7f,
+    static constexpr float undo_redo_width  = height * 0.7f,
                          undo_redo_height = undo_redo_width;
     
-    static constexpr int change_preset_width  = undo_redo_width,
+    static constexpr float change_preset_width  = undo_redo_width,
                          change_preset_height = undo_redo_width;
     
-    static constexpr int preset_display_width  = 100,
+    static constexpr float preset_display_width  = 100,
                          preset_display_height = undo_redo_width;
     
-    static constexpr int A_B_Copy_width  = undo_redo_width,
+    static constexpr float A_B_Copy_width  = undo_redo_width,
                          A_B_Copy_height = undo_redo_width;
 };
 
@@ -64,7 +64,7 @@ public:
 class MainPanelGUI {
 public:
     //==============================================================================
-    static constexpr int width  = PluginNameGUI::width - MenuPanelGUI::width,
+    static constexpr float width  = PluginNameGUI::width - MenuPanelGUI::width,
                          height = PluginNameGUI::height - PresetPanelGUI::height;
     
 };
@@ -73,7 +73,7 @@ public:
 class SidePanelGUI {
 public:
     //==============================================================================
-    static constexpr int width  = PluginNameGUI::width - PluginNameGUI::height,
+    static constexpr float width  = PluginNameGUI::width - PluginNameGUI::height,
                          height = PluginNameGUI::height;
     
 };
@@ -82,7 +82,7 @@ public:
 class DisplayPresetOverlayGUI {
 public:
     //==============================================================================
-    static constexpr int width  = PluginNameGUI::width - MenuPanelGUI::width,
+    static constexpr float width  = PluginNameGUI::width - MenuPanelGUI::width,
                          height = PluginNameGUI::height;
 };
 
@@ -90,7 +90,7 @@ public:
 class ActionPresetOverlayGUI {
 public:
     //==============================================================================
-    static constexpr int width  = PluginNameGUI::width,
+    static constexpr float width  = PluginNameGUI::width,
                          height = PluginNameGUI::height;
 };
